+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added error code to on office exception
 - Added a method to get the error from the on office exception
 - Added a stringify method to an office error
+- Added a handling for response errors in the status key
 
 ## v0.9.1
 - Added the api claim back to the request factory